♻️ refactor(cli): move commit command logic to cli/commit.rs
- relocate Commit struct and methods to cli/commit.rs for better organization
- update run_commit function to be a method of Commit struct
- adjust main.rs to use the new method approach for commit command
