+++ src/cli/commit.rs
@@ -1,20 +1,48 @@
 use crate::Sign;
 
+use clap::Parser;
 use color_eyre::Result;
 
-use super::{CIExit, Commands, Commit};
+use super::{CIExit, Commands};
 
-pub async fn run_commit(sign: Sign, args: Commit) -> Result<CIExit> {
-    let client = super::get_client(Commands::Commit(args.clone())).await?;
+/// Configuration for the Commit command
+#[derive(Debug, Parser, Clone)]
+pub struct Commit {
+    /// Semantic version number for a tag
+    #[arg(short, long)]
+    pub semver: Option<String>,
+    /// Message to add to the commit when pushing
+    #[arg(short, long)]
+    commit_message: String,
+    /// Prefix for the version tag
+    #[clap(short, long, default_value_t = String::from("v"))]
+    pub prefix: String,
+}
 
-    super::commit_changed_files(
-        &client,
-        sign,
-        args.commit_message(),
-        &args.prefix,
-        args.tag_opt(),
-    )
-    .await?;
+impl Commit {
+    pub fn commit_message(&self) -> &str {
+        &self.commit_message
+    }
 
-    Ok(CIExit::Committed)
+    pub fn tag_opt(&self) -> Option<&str> {
+        if let Some(semver) = &self.semver {
+            return Some(semver);
+        }
+        None
+    }
+
+    pub async fn run_commit(&self, sign: Sign) -> Result<CIExit> {
+        let client = super::get_client(Commands::Commit(self.clone())).await?;
+
+        super::commit_changed_files(
+            &client,
+            sign,
+            self.commit_message(),
+            &self.prefix,
+            self.tag_opt(),
+        )
+        .await?;
+
+        Ok(CIExit::Committed)
+    }
 }
